Add files via upload
## Problem 1 : Longest Stable Subsequence

Consider a list of numbers $[a_0, a_1, \cdots, a_{n-1}]$. Our goal is to find the the longest stable subsequence: $[a_{i_1}, a_{i_2}, \cdots, a_{i_k} ]$ which is a sub-list of the original list that selects elements at indices $i_1, i_2, \ldots, i_k$ from the original list such that 
  1. $i_1 < i_2 < \cdots < i_k$; 
  2. $a_{i_{j+1}} - 1 \leq a_{i_{j}} \leq a_{i_{j+1}} + 1 $. We can also write this as $|a_{i_{j+1}} - a_{i_j}| \leq 1$. I.e, each element of the subsequence must be within $\pm 1$ or equal to the previous element.
  3. The length of the subsequence $k$ is maximized.

### Example 

Consider the list `[1, 4, 2, -2, 0, -1, 2, 3]`.  There are many "stable subsequences":
 - `[1, 0, -1]`
 - `[1, 2, 2, 3]`
 - `[4, 3]`

The longest stable subsequence is `[1, 2, 2, 3]` of length 4. Note that each element of the subsequence is at most $1$ away from the previous element.

The goal of this problem is to formulate a dynamic programming solution to find the length of the longest stable subsequence and the subsequence itself.